Trait rabbitmq_adaptor::RabbitExt

pub trait RabbitExt {
    fn create_topic_exchange(
        self, 
        name: &'static str
    ) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>;

    fn create_queue(
        self, 
        name: String
    ) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>;

    fn bind_queue(
        self, 
        name: &'static str, 
        exchange: &'static str, 
        routing_key: &'static str
    ) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>;

    fn register_consumer<F: Fn(Delivery) -> bool + Send + 'static>(
        self, 
        queue_name: String, 
        callback: F
    ) -> Box<dyn Future<Item = (), Error = FailureError> + Send>;

    fn publish(
        self, 
        exchange: &'static str, 
        topic: &'static str, 
        payload: Vec<u8>
    ) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>;

    fn rpc_publish(
        self, 
        exchange: &'static str, 
        topic: &'static str, 
        payload: Vec<u8>
    ) -> Box<dyn Future<Item = Delivery, Error = FailureError> + Send>;

    fn rpc_reply(
        self, 
        message: &Delivery, 
        payload: Vec<u8>
    ) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>;

    fn close_channel(
        self
    ) -> Box<dyn Future<Item = (), Error = FailureError> + Send>;
}

RabbitExt overloads the futures generated from the connect() method to allow for easy chaining of futures.

Required methods

fn create_topic_exchange(
self,
name: &'static str
) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>

Creates a topic exchange https://www.rabbitmq.com/tutorials/amqp-concepts.html#exchange-topic

fn create_queue(
self,
name: String
) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>

Creates a queue https://www.rabbitmq.com/tutorials/amqp-concepts.html#queues

fn bind_queue(
self,
name: &'static str,
exchange: &'static str,
routing_key: &'static str
) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>

Binds a queue to a topic exchange https://www.rabbitmq.com/tutorials/amqp-concepts.html#bindings

fn register_consumer<F: Fn(Delivery) -> bool + Send + 'static>(
self,
queue_name: String,
callback: F
) -> Box<dyn Future<Item = (), Error = FailureError> + Send>

Register a push consumer, supplying a callback which is executed on each incoming message The callback function should return true for ack and false for nack https://www.rabbitmq.com/tutorials/amqp-concepts.html#consumers

fn publish(
self,
exchange: &'static str,
topic: &'static str,
payload: Vec<u8>
) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>

Publish a message to a topic exchange

fn rpc_publish(
self,
exchange: &'static str,
topic: &'static str,
payload: Vec<u8>
) -> Box<dyn Future<Item = Delivery, Error = FailureError> + Send>

Publish a RPC message to a topic exchange

fn rpc_reply(
self,
message: &Delivery,
payload: Vec<u8>
) -> Box<dyn Future<Item = RabbitConnection, Error = FailureError> + Send>

fn close_channel(
self
) -> Box<dyn Future<Item = (), Error = FailureError> + Send>

Close the channel gracefully

Loading content...